Bety/requirements.txt
# ============================================================
#  Privacy-First Local Voice Assistant — requirements.txt
# ============================================================

# UI framework
customtkinter>=5.2.0

# Whisper speech-to-text (fast CPU-friendly implementation)
faster-whisper>=1.0.0

# Audio capture
sounddevice>=0.4.6
numpy>=1.24.0

# WAV encoding (writing audio buffer to in-memory file)
scipy>=1.11.0

# Global dictation mode — hotkey listener + virtual keyboard input
keyboard>=0.13.5
pyautogui>=0.9.54

# System tray integration
pystray>=0.19.4
Pillow>=10.0.0

# Packaging — compile to a standalone .exe via build.py
pyinstaller>=6.0.0
